Fix a broken build option "--with-gl-provider"
In the previous code, the configure option "--with-gl-provider" with a
supported provider name (such as EGL) causes a compile error because
the specified GLContextProvider is declared twice. In addition it's
also meaningless for unknown or unsupported providers because the
implementation for it doesn't exist in this case, so it causes a link
error.

Signed-off-by: Takuro Ashie <ashie@homa.ne.jp>
Signed-off-by: Hiroshi Hatake <hatake@clear-code.com>